+++ packages/make-styles/src/makeStyles.test.ts
@@ -28,7 +28,7 @@ describe('makeStyles', () => {
         color: 'red',
       },
     });
-    expect(computeClasses({ dir: 'ltr', renderer, tokens: {} }).root).toEqual('__ncdyee0 fe3e8s90');
+    expect(computeClasses({ dir: 'ltr', renderer }).root).toEqual('__ncdyee0 fe3e8s90');
 
     expect(renderer).toMatchInlineSnapshot(`
       .fe3e8s90 {
@@ -44,7 +44,7 @@ describe('makeStyles', () => {
         position: 'absolute',
       },
     });
-    expect(computeClasses({ dir: 'ltr', renderer, tokens: {} }).root).toEqual('__1fslksb fe3e8s90 f1euv43f');
+    expect(computeClasses({ dir: 'ltr', renderer }).root).toEqual('__1fslksb fe3e8s90 f1euv43f');
 
     expect(renderer).toMatchInlineSnapshot(`
       .fe3e8s90 {
@@ -64,8 +64,8 @@ describe('makeStyles', () => {
       },
     });
 
-    const ltrClasses = computeClasses({ dir: 'ltr', renderer, tokens: {} }).root;
-    const rtlClasses = computeClasses({ dir: 'rtl', renderer, tokens: {} }).root;
+    const ltrClasses = computeClasses({ dir: 'ltr', renderer }).root;
+    const rtlClasses = computeClasses({ dir: 'rtl', renderer }).root;
 
     expect(ltrClasses).toEqual('__947mlk0 frdkuqy0 f1c8chgj');
     expect(rtlClasses).toEqual('__hcjvlo0 rfrdkuqy0 rf1c8chgj');
@@ -101,7 +101,7 @@ describe('makeStyles', () => {
         animationDuration: '5s',
       },
     });
-    expect(computeClasses({ dir: 'rtl', renderer, tokens: {} }).root).toBe('__194gjlt rf1g6ul6r f1cpbl36 f1t9cprh');
+    expect(computeClasses({ dir: 'rtl', renderer }).root).toBe('__194gjlt rf1g6ul6r f1cpbl36 f1t9cprh');
 
     expect(renderer).toMatchInlineSnapshot(`
       @-webkit-keyframes rf1q8eu9e {
@@ -141,10 +141,10 @@ describe('makeStyles', () => {
       root: { display: 'flex', paddingLeft: '10px' },
     });
 
-    const classesA = computeClasses({ dir: 'rtl', renderer: rendererA, tokens: {} }).root;
+    const classesA = computeClasses({ dir: 'rtl', renderer: rendererA }).root;
 
-    computeClasses({ dir: 'ltr', renderer: rendererB, tokens: {} }).root;
-    const classesB = computeClasses({ dir: 'rtl', renderer: rendererB, tokens: {} }).root;
+    computeClasses({ dir: 'ltr', renderer: rendererB }).root;
+    const classesB = computeClasses({ dir: 'rtl', renderer: rendererB }).root;
 
     // Classes emitted by different renderers can be the same
     expect(classesA).toBe(classesB);
@@ -168,6 +168,21 @@ describe('makeStyles', () => {
       }
       .rfrdkuqy0 {
         padding-right: 10px;
+      }
+    `);
+  });
+  it('handles tokens', () => {
+    const rendererA = createDOMRenderer(createFakeDocument());
+
+    const computeClasses = makeStyles<'root', { display: string }>({
+      root: tokens => ({ display: tokens.display }),
+    });
+
+    computeClasses({ dir: 'rtl', renderer: rendererA });
+
+    expect(rendererA).toMatchInlineSnapshot(`
+      .fl81ese0 {
+        display: var(--display);
       }
     `);
   });